feat: 공개 범위 저장 시 낙관적 업데이트 적용
