+++ src/app/setting/components/SettingPrivacy.tsx
@@ -14,8 +14,10 @@ type Props = {
 const testUserId = "588a4dea-b95a-4836-b6bc-10dbafa4a81f";
 
 const SettingPrivacy = ({ setting }: Props) => {
-  const { privacySelected, handlePrivacySelectedChange, isButtonEnabled } = useSettingPrivacy(setting);
-  const mutatePatchPrivacy = usePatchPrivacy();
+  const { privacySelected, handlePrivacySelectedChange, isButtonEnabled, handlePatchPrivacy } =
+    useSettingPrivacy(setting);
+  // TODO: Mutate 되면 버튼 조건도 바뀌어야함 -> isButtonEnabled
+  const mutatePatchPrivacy = handlePatchPrivacy();
 
   console.log("privacySelected", privacySelected);
   console.log("isButtonEnabled", isButtonEnabled);
@@ -27,6 +29,7 @@ const SettingPrivacy = ({ setting }: Props) => {
         {isButtonEnabled && (
           <button
             className="bg-slate-200"
+            // onClick={() => mutatePatchPrivacy({ userId: testUserId, privacyType: privacySelected })}>
             onClick={() => mutatePatchPrivacy({ userId: testUserId, privacyType: privacySelected })}>
             저장
           </button>
